refactor : 잔액 차감 검증을 포인트 내부의 Balance 객체에게 위임
