--- src/main/java/kr/hhplus/be/server/domain/point/pointhistory/PointHistoryRepository.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.domain.point.pointhistory;
-
-public interface PointHistoryRepository {
-
-	PointHistory save(PointHistory pointHistory);
-
-}
